+++ vault/logical_system_paths.go
@@ -708,17 +708,13 @@ func (b *SystemBackend) pluginsReloadPath() *framework.Path {
 				Type:        framework.TypeCommaStringSlice,
 				Description: strings.TrimSpace(sysHelp["plugin-backend-reload-mounts"][0]),
 			},
-			"scope": &framework.FieldSchema{
-				Type:        framework.TypeString,
-				Description: strings.TrimSpace(sysHelp["plugin-backend-reload-scope"][0]),
-			},
 		},
 
 		Operations: map[logical.Operation]framework.OperationHandler{
 			logical.UpdateOperation: &framework.PathOperation{
 				Callback:    b.handlePluginReloadUpdate,
 				Summary:     "Reload mounted plugin backends.",
-				Description: "Either the plugin name (`plugin`) or the desired plugin backend mounts (`mounts`) must be provided, but not both. In the case that the plugin name is provided, all mounted paths that use that plugin backend will be reloaded.  If (`scope`) is provided and is (`cluster`), the plugin(s) are reloaded cluster wide.",
+				Description: "Either the plugin name (`plugin`) or the desired plugin backend mounts (`mounts`) must be provided, but not both. In the case that the plugin name is provided, all mounted paths that use that plugin backend will be reloaded.",
 			},
 		},
 
@@ -726,6 +722,7 @@ func (b *SystemBackend) pluginsReloadPath() *framework.Path {
 		HelpDescription: strings.TrimSpace(sysHelp["plugin-reload"][1]),
 	}
 }
+
 func (b *SystemBackend) toolsPaths() []*framework.Path {
 	return []*framework.Path{
 		{